feat(@clayui/core): adds new API to control node expand state in TreeView
Exposes a public API that is exposed via render props to allow the developer to call according to their use case.
