+++ packages/clay-core/src/tree-view/Collection.tsx
@@ -3,19 +3,15 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import React, {Key, useCallback} from 'react';
+import React from 'react';
 
-import {useTreeViewContext} from './context';
+import {Expand, Selection, useAPI} from './context';
 import {ItemContextProvider, useItem} from './useItem';
-
-export type Selection = {
-	toggle: (key: Key) => void;
-	has: (key: Key) => boolean;
-};
 
 export type ChildrenFunction<T> = (
 	item: Omit<T, 'indexes' | 'itemRef' | 'key' | 'parentItemRef'>,
-	selection: Selection
+	selection: Selection,
+	expand: Expand
 ) => React.ReactElement;
 
 export interface ICollectionProps<T> {
@@ -58,24 +54,17 @@ export function Collection<T extends Record<any, any>>({
 	children,
 	items,
 }: ICollectionProps<T>) {
-	const {selection} = useTreeViewContext();
+	const api = useAPI();
 	const {key: parentKey} = useItem();
-
-	const hasKey = useCallback(
-		(key: Key) => {
-			return selection.selectedKeys.has(key);
-		},
-		[selection.selectedKeys]
-	);
 
 	return (
 		<>
 			{typeof children === 'function' && items
 				? items.map((item, index) => {
-						const child = children(removeItemInternalProps(item), {
-							has: hasKey,
-							toggle: selection.toggleSelection,
-						});
+						const child = children(
+							removeItemInternalProps(item),
+							...api
+						);
 
 						const key = getKey(
 							index,
