+++ packages/clay-core/src/tree-view/context.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import React, {useContext} from 'react';
+import React, {Key, useCallback, useContext} from 'react';
 
 import type {ChildrenFunction} from './Collection';
 import type {ITreeState} from './useTree';
@@ -35,4 +35,38 @@ export const TreeViewContext = React.createContext<ITreeViewContext<any>>(
 
 export function useTreeViewContext(): ITreeViewContext<unknown> {
 	return useContext(TreeViewContext);
+}
+
+export type Selection = {
+	toggle: (key: Key) => void;
+	has: (key: Key) => boolean;
+};
+
+export type Expand = {
+	toggle: (key: Key) => void;
+	has: (key: Key) => boolean;
+};
+
+export function useAPI() {
+	const {expandedKeys, selection, toggle} = useTreeViewContext();
+
+	const hasKey = useCallback(
+		(key: Key) => {
+			return selection.selectedKeys.has(key);
+		},
+		[selection.selectedKeys]
+	);
+
+	const hasExpandedKey = useCallback(
+		(key: Key) => expandedKeys.has(key),
+		[expandedKeys]
+	);
+
+	return [
+		{
+			has: hasKey,
+			toggle: selection.toggleSelection,
+		},
+		{has: hasExpandedKey, toggle},
+	] as const;
 }
